refactoring sum testing for reuse
