+++ test/summation/sum_test.dart
@@ -13,95 +13,7 @@ import 'package:rohd/rohd.dart';
 import 'package:rohd_hcl/rohd_hcl.dart';
 import 'package:test/test.dart';
 
-int goldenSumOfLogics(
-  List<Logic> logics, {
-  required int width,
-  bool saturates = false,
-  int? maxVal,
-  int minVal = 0,
-}) =>
-    goldenSum(
-      logics.map((e) => SumInterface(width: e.width)..amount.gets(e)).toList(),
-      width: width,
-      saturates: saturates,
-      minVal: minVal,
-      maxVal: maxVal,
-    );
-
-int goldenSum(
-  List<SumInterface> interfaces, {
-  required int width,
-  bool saturates = false,
-  int? maxVal,
-  int? minVal,
-  int initialValue = 0,
-  bool debug = false,
-}) {
-  void log(String message) {
-    if (debug) {
-      // ignore: avoid_print
-      print(message);
-    }
-  }
-
-  log('width: $width');
-
-  var sum = initialValue;
-
-  log('min $minVal  ->  max $maxVal');
-
-  maxVal ??= (1 << width) - 1;
-  if (maxVal > (1 << width) - 1) {
-    // ignore: parameter_assignments
-    maxVal = (1 << width) - 1;
-  }
-  minVal ??= 0;
-
-  log('min $minVal  ->  max $maxVal  [adjusted]');
-
-  if (minVal > maxVal) {
-    throw Exception('minVal must be less than or equal to maxVal');
-  }
-
-  log('init: $initialValue');
-
-  for (final intf in interfaces) {
-    final amount = intf.amount.value.toInt();
-    final enabled = !intf.hasEnable || intf.enable!.value.toBool();
-
-    log('${intf.increments ? '+' : '-'}'
-        '$amount${enabled ? '' : '  [disabled]'}');
-
-    if (enabled) {
-      if (intf.increments) {
-        sum += amount;
-      } else {
-        sum -= amount;
-      }
-    }
-  }
-
-  log('=$sum');
-
-  if (saturates) {
-    if (sum > maxVal) {
-      sum = maxVal;
-    } else if (sum < minVal) {
-      sum = minVal;
-    }
-    log('saturates to $sum');
-  } else {
-    final range = maxVal - minVal + 1;
-    if (sum > maxVal) {
-      sum = (sum - maxVal - 1) % range + minVal;
-    } else if (sum < minVal) {
-      sum = maxVal - (minVal - sum - 1) % range;
-    }
-    log('rolls-over to $sum');
-  }
-
-  return sum;
-}
+import 'summation_test_utils.dart';
 
 void main() {
   test('simple sum of 1 ofLogics', () async {
@@ -372,53 +284,10 @@ void main() {
   test('random', () {
     final rand = Random(123);
 
-    SumInterface genRandomInterface() {
-      final isFixed = rand.nextBool();
-      return SumInterface(
-        fixedAmount: isFixed ? rand.nextInt(100) : null,
-        width: isFixed ? null : rand.nextInt(8),
-        increments: rand.nextBool(),
-        hasEnable: rand.nextBool(),
-      );
-    }
-
-    List<SumInterface> genRandomInterfaces() {
-      final numInterfaces = rand.nextInt(8) + 1;
-      return List.generate(numInterfaces, (_) => genRandomInterface());
-    }
-
     for (var i = 0; i < 1000; i++) {
-      final interfaces = genRandomInterfaces();
+      final cfg = genRandomSummationConfiguration();
 
-      final width = rand.nextBool() ? null : rand.nextInt(10) + 1;
-
-      final saturates = rand.nextBool();
-      var minVal = rand.nextBool() ? rand.nextInt(30) : 0;
-      var maxVal = rand.nextBool()
-          ? rand.nextInt(width == null ? 70 : ((1 << width) - 1)) + minVal + 1
-          : null;
-      var initialValue = rand.nextBool() ? rand.nextInt(maxVal ?? 100) : 0;
-
-      if (maxVal != null && width != null) {
-        // truncate to width
-        maxVal = max(1, LogicValue.ofInt(maxVal, width).toInt());
-      }
-
-      if (width != null) {
-        // truncate to width
-        initialValue = LogicValue.ofInt(initialValue, width).toInt();
-      }
-
-      if (maxVal == null || minVal >= maxVal) {
-        if (maxVal == null && width == null) {
-          minVal = 0;
-        } else {
-          minVal =
-              rand.nextInt(maxVal ?? (width == null ? 0 : (1 << width) - 1));
-        }
-      }
-
-      for (final intf in interfaces) {
+      for (final intf in cfg.interfaces) {
         if (intf.hasEnable) {
           intf.enable!.put(rand.nextBool());
         }
@@ -428,37 +297,21 @@ void main() {
         }
       }
 
-      int safeWidthFor(int val) {
-        final lv = LogicValue.ofInferWidth(val);
-        final inferredWidth = lv.width;
-
-        return min(max(inferredWidth, 1), width ?? inferredWidth);
-      }
-
-      final dut = Sum(interfaces,
-          saturates: saturates,
-          maxValue: maxVal != null && rand.nextBool()
-              ? Const(LogicValue.ofInferWidth(maxVal),
-                  width: safeWidthFor(maxVal))
-              : maxVal,
-          minValue: rand.nextBool()
-              ? Const(LogicValue.ofInferWidth(minVal),
-                  width: safeWidthFor(minVal))
-              : minVal,
-          width: width,
-          initialValue: rand.nextBool()
-              ? Const(LogicValue.ofInferWidth(initialValue),
-                  width: safeWidthFor(initialValue))
-              : initialValue);
+      final dut = Sum(cfg.interfaces,
+          saturates: cfg.saturates,
+          maxValue: cfg.maxValue,
+          minValue: cfg.minValue,
+          width: cfg.width,
+          initialValue: cfg.initialValue);
 
       final actual = dut.sum.value.toInt();
       final expected = goldenSum(
-        interfaces,
+        cfg.interfaces,
         width: dut.width,
-        saturates: saturates,
-        maxVal: maxVal,
-        minVal: minVal,
-        initialValue: initialValue,
+        saturates: cfg.saturates,
+        maxVal: cfg.maxVal,
+        minVal: cfg.minVal,
+        initialValue: cfg.initialVal,
       );
 
       expect(actual, expected);
